pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [
    "Michael Asumadu <michaelasumadu1@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.8"
pip = "21.1.2"
Flask = "2.2.2"
kafka-python = "^2.0.2"
psycopg2-binary = "^2.8.6"
Flask-Migrate = "3.1.0"
requests = "^2.25.1"
python-dotenv = "0.21.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "0.28.1"
Flask-Script = "^2.0.6"
loguru = "0.6.0"
flask-swagger-ui = "4.11.1"
pinject = "^0.14.1"
apispec = "5.2.2"
apispec-webframeworks = "^0.5.2"
flask-mongoengine = "^1.0.0"
redis = "4.3.4"
hiredis = "^2.0.0"
Faker = "14.2.1"
fakeredis = "^1.5.0"
marshmallow-enum = "1.5.1"
blinker = "^1.4"
pytz = "2022.2.1"
PyJWT = "^2.1.0"
gunicorn = "^20.1.0"
virtualenv = "^20.14.0"
Flask-Cors = "^3.0.10"
marshmallow = "3.18.0"
Flask-AutoIndex = "^0.6.6"
cryptography = "38.0.1"
Pillow = "^9.1.0"
boto3 = "^1.21.46"
gevent = "^21.12.0"
pylint = "^2.13.9"
locust = "^2.10.1"
py-healthcheck = "^1.10.1"
tomli = "2.0.1"
tox = "^4.4.4"

[tool.poetry.dev-dependencies]
pytest = "7.1.3"
black = "22.8.0"
flake8 = "5.0.4"
pre-commit = "^2.12.1"
Flask-Testing = "^0.8.1"
mongomock = "4.1.2"
fakeredis = "^1.5.0"
pytest-cov = "^3.0.0"

[tool.poetry.group.dev.dependencies]
mongomock = "^4.1.2"

[tool.black]
line-length = 89
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(/(
    \.git
  | \.hg
  | \.eggs
  | \.ini
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .yml
  | .yaml
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
